Refactor the observation broadener code to allow broadeners to maintain state.  Increase test coverage.
