+++ src/main/scala/supervisedLearning/linearRegression/observationBroadener/AddRaisedDimension.scala
@@ -1,29 +1,22 @@
 package supervisedLearning.linearRegression.observationBroadener
 import breeze.linalg.DenseVector
 
-object AddRaisedDimension extends CreatesObservationDimension {
+class AddRaisedDimension(override val index: Int, private val power: Double) extends ExtendObservationDimensions(index) {
 	
 	/**
-	  * Given an index, an observation and a power create a new observation that adds a new dimension to the given
-	  * observation by taking its dimension, retrieved using the index, and raising it to the given power.
-	  * 
-	  * @param index a 0-based index used to determine which dimension of the observation should be used as a basis for
-	  *              the new dimension
+	  * Given an observation, extend it by adding a new dimension generated by taking the dimension indexed by index
+	  * and raising it by a power.
+	  *
 	  * @param observation an observation to which we would like to add a dimension
-	  * @param power the power to which we would like to raise a given dimension to create a new dimension
-	  * @return an observation
+	  * @return the extended observation
 	  */
-	override def getExtendedObservation(index: Int, observation: DenseVector[Double], power:Double*): DenseVector[Double] = {
+	override def getExtendedObservation(observation: DenseVector[Double]): DenseVector[Double] = {
 		
-		if (super.isIndexOutOfBounds(index, observation)) {
+		if (super.isIndexOutOfBounds(observation)) {
 			return observation
 		}
 		
-		if (!power.length.equals(1)) {
-			return observation
-		}
-		
-		val newDimensionValue = scala.math.pow(observation(index), power(0))
+		val newDimensionValue = scala.math.pow(observation(index), power)
 		super.extendObservation(observation, newDimensionValue)
 	}
 }
